requirements.txt
paho-mqtt>=1.6.1
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0